pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aievodev"
version = "0.1.0"
description = "Self-optimizing AI test generator using adversarial evolution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0,<3.0.0",
    "google-genai>=1.0.0",
    "langchain>=1.0.0,<2.0.0",
    "langchain-classic>=1.0.0,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    "langchain-google-genai>=2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
    "pytest>=8.0.0,<10.0.0",
    "coverage>=7.0.0,<8.0.0",
]

[project.optional-dependencies]
dev = [
    "tiktoken>=0.5.0,<1.0.0",
]

[project.scripts]
aievodev = "src.core.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*", "specs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
